Simplify: shared authed-download helper, hoisted origin gate
Cleanup pass over the import feature (reuse/simplification review):

- One downloadFromApi helper in dashboard-core/lib replaces the near-
  identical blob-download implementations in use-export and use-import.
  The import copy had dropped the deliberate same-origin gate that keeps
  the JWT off third-party hosts — the shared helper restores it for all
  callers, and no longer touches window at module scope
- validated_results_url, byte-identical in the imports and exports
  controllers, hoists to their shared Admin ResourceController as
  validated_allowed_origin_url; the redundant allowed_origins.exists?
  pre-check goes away (allowed_origin? is already false on an empty
  allow-list)
- The mapping table's not-mapped option now carries disabled: false so
  the render drops its 'disabled' in option type guard
- ImportButton stops overriding the browse label with a string identical
  to the FileUploadField default; the duplicated key leaves all six
  locales
