+++ packages/dashboard-core/src/hooks/use-import.ts
@@ -1,10 +1,9 @@
 import type { Import, ImportCreateParams, ImportType } from '@spree/admin-sdk'
 import { useMutation } from '@tanstack/react-query'
 import { adminClient } from '../client'
+import { downloadFromApi } from '../lib/download'
 import { useStore } from '../providers/store-provider'
 import { useAuth } from './use-auth'
-
-const API_BASE_URL = import.meta.env.VITE_SPREE_API_URL || ''
 
 export interface CreateImportInput {
   type: ImportType
@@ -32,31 +31,6 @@ export function useCreateImport() {
         results_url: `${window.location.origin}/${storeId}/settings/imports`,
       }),
   })
-}
-
-/**
- * Authed fetch → Blob URL download. The endpoints are JWT-protected, so a
- * top-level navigation (which cannot carry the in-memory token) does not work.
- */
-async function downloadFromApi(token: string | null, path: string, fallbackName: string) {
-  const headers: Record<string, string> = {}
-  if (token) headers.Authorization = `Bearer ${token}`
-
-  const response = await fetch(`${API_BASE_URL}${path}`, { headers, credentials: 'include' })
-  if (!response.ok) throw new Error(`Download failed: ${response.status}`)
-
-  const disposition = response.headers.get('Content-Disposition')
-  const filename = disposition?.match(/filename="?([^";]+)"?/)?.[1] ?? fallbackName
-
-  const blob = await response.blob()
-  const objectUrl = URL.createObjectURL(blob)
-  const anchor = document.createElement('a')
-  anchor.href = objectUrl
-  anchor.download = filename
-  document.body.appendChild(anchor)
-  anchor.click()
-  anchor.remove()
-  URL.revokeObjectURL(objectUrl)
 }
 
 /** Downloads the CSV template for an import type. */
