+++ packages/dashboard-core/src/locales/ar.json
@@ -481,7 +481,6 @@
         "title": "استيراد من CSV",
         "description": "قم برفع ملف CSV. ستقوم بمطابقة أعمدته مع الحقول في الخطوة التالية.",
         "drop_label": "أسقط ملف CSV هنا، أو انقر للاستعراض",
-        "browse": "استعراض الملفات",
         "delimiter_label": "فاصل الأعمدة",
         "delimiters": {
           "comma": "فاصلة (,)",
